Moved carousel progress to new file placement, rfce prepared for all bottom row jsx files - NS
