+++ client/src/components/game_layout/bottom_row/dice/Dice.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Dice() {
+  return (
+    <>
+      <h1>Dice</h1>
+    </>
+  );
+}
+
+export default Dice;
