+++ client/src/components/game_layout/bottom_row/dice/Die.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Die() {
+  return (
+    <>
+      <h1>Die</h1>
+    </>
+  );
+}
+
+export default Die;
