+++ client/src/components/game_layout/bottom_row/revealed_card/RevealedCard.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function RevealedCard() {
+  return (
+    <>
+      <h1>RevealedCard</h1>
+    </>
+  );
+}
+
+export default RevealedCard;
